Typescript implementation - All implemented with pending testimonials & trending components
