--- src/components/logo/logo.component.jsx
@@ -1,14 +1,0 @@
-import { LogoContainer, TitleContainer, TitleOne, TitleTwo } from "./logo.styles";
-
-const Logo = () => {
-  return (
-    <LogoContainer to='/'>
-      <TitleContainer>
-        <TitleOne>FABRIC</TitleOne>
-        <TitleTwo>CLOTHING</TitleTwo>
-      </TitleContainer>
-    </LogoContainer>
-  )
-}
-
-export default Logo;
