+++ src/components/logo/logo.styles.tsx
@@ -1,0 +1,27 @@
+import styled from "styled-components";
+
+export const TitleContainer = styled.div`
+	padding: 0px;
+	margin: auto 0;
+	font-size: 1.25rem;
+	font-family: "Bebas Neue", sans-serif;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	white-space: nowrap;
+	background-color: white;
+
+	&:hover {
+		cursor: pointer;
+	}
+`;
+
+export const TitleOne = styled.span`
+	padding: 0.1em 0.25em 0.08em 0.25em;
+	background-color: black;
+	color: white;
+`;
+
+export const TitleTwo = styled.span`
+	padding: 0.1em 0.25em 0.08em 0.25em;
+`;
