--- src/components/release/release.component.jsx
@@ -1,71 +1,0 @@
-import { FaLink } from "react-icons/fa";
-import { ReleaseContainer, ReleaseDesc, ReleaseTitle } from './release.styles';
-
-const Release = () => {
-  return (
-    <>
-      <ReleaseContainer>
-        <ReleaseTitle>
-          Release <i>v </i>1.0.0 <a
-            href="https://github.com/Nidhish-D-Nayak/fabric-clothing"
-            target="_blank"
-            rel="noreferrer noopener">
-            <FaLink />
-          </a>
-        </ReleaseTitle>
-        <ReleaseDesc>
-          <li>
-            Redux by @Nidhish-D-Nayak in <a
-              href="https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/4"
-              target="_blank"
-              rel="noreferrer noopener">
-              https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/4
-            </a>
-          </li>
-          <li>
-            Redux Extended Tools by @Nidhish-D-Nayak in <a
-              href="https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/7"
-              target="_blank"
-              rel="noreferrer noopener">
-              https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/7
-            </a>
-          </li>
-          <li>
-            Redux Saga by @Nidhish-D-Nayak in <a
-              href="https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/8"
-              target="_blank"
-              rel="noreferrer noopener">
-              https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/8
-            </a>
-          </li>
-          <li>
-            Razorpay Integration feature by @Nidhish-D-Nayak in <a
-              href="https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/11"
-              target="_blank"
-              rel="noreferrer noopener">
-              https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/11
-            </a>
-          </li>
-          <li>
-            Redux Toolkit by @Nidhish-D-Nayak in <a
-              href="https://github.com/Nidhish-D-Nayak/fabric-clothing/commit/8011d386272d6c489a25a6ae81b0ede12fa3eb10"
-              target="_blank"
-              rel="noreferrer noopener">
-              https://github.com/Nidhish-D-Nayak/fabric-clothing/commit/8011d386272d6c489a25a6ae81b0ede12fa3eb10
-            </a>
-          </li>
-          <li>
-            UI Overhaul by @Nidhish-D-Nayak in <a
-              href="https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/12"
-              target="_blank"
-              rel="noreferrer noopener">
-              https://github.com/Nidhish-D-Nayak/fabric-clothing/pull/12
-            </a>
-          </li>
-        </ReleaseDesc>
-      </ ReleaseContainer>
-    </>
-  )
-}
-
-export default Release;
