--- src/components/spinner/spinner.component.jsx
@@ -1,11 +1,0 @@
-import { SpinnerContainer, SpinnerOverlay } from './spinner.styles';
-
-const Spinner = () => {
-    return (
-        <SpinnerOverlay>
-            <SpinnerContainer />
-        </SpinnerOverlay>
-    )
-}
-
-export default Spinner;
