+++ src/components/spinner/spinner.component.tsx
@@ -1,0 +1,11 @@
+import { SpinnerContainer, SpinnerOverlay } from "./spinner.styles";
+
+const Spinner = () => {
+	return (
+		<SpinnerOverlay>
+			<SpinnerContainer />
+		</SpinnerOverlay>
+	);
+};
+
+export default Spinner;
